Use newly created mobdev_err.h
The new error translation is no longer sourced by copying it out of other projects with debatable origin.
It is created by https://github.com/nanoscopic/mobdev_dump_errs
Ultimately the data is extracted from the MobileDevice framework itself.
Note that it isn't actually necessary to copy this information out from the framework and it could be referenced in the same way it is extracted to begin with.
This will be done, as I don't really like pointlessly copying this information.
If it is used from the source instead then it will also in theory be translated into other languages also ( at least if other translations of MacOS include a version of MobileDevice with these errors translated... )