Many updates
Basically everything has been changed.

A major change is that Apple NSKeyedArchiver is no longer used to created NSKeyedArchives, instead I have created an entirely new implementation in C. The purpose of this is to avoid having to have any Objective C in the codebase, and to not have to link against the Foundation library. This is not yet complete; NSKeyedUnarchiving will need to be implemented as well.

Many new commands have been added, and the entire codebase has been heavily refactored and cleaned up.