Minor tweaks
Added check for blank messages on iOS 11, as they occur on some versions and were crashing xctest.

Added back the guard for small message unarchiving that are too small to contain anything. It shouldn't ever occur but is a good sanity.

Removed initial call on xctest for iOS 11 that has a blank pid. This call doesn't exist on some iOS 11 versions ( before 11.3? ) and so was crashing on those versions. It isn't even needed for xctest to work so commented it out for now.